B11__B20_Projects/B20_Readify_20_project/backend/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==2.0.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
cssselect2==0.8.0
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.24
django-allauth==65.11.2
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
environs==14.3.0
exceptiongroup==1.3.0
fonttools==4.59.2
frozenlist==1.7.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
magic-filter==1.0.12
marshmallow==4.0.1
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
phonenumberslite==9.0.15
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.10.0
yarl==1.20.1
zopfli==0.2.3.post1
